requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
requests>=2.32.0
python-dateutil>=2.9.0
scipy>=1.13.0
openpyxl>=3.1.5
tencentcloud-sdk-python>=3.0.1280
python-dotenv>=1.0.1
